Added unit tests for service
